+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/BUILD.bazel
@@ -30,6 +30,7 @@ iree_gentbl_cc_library(
 iree_compiler_cc_library(
     name = "Transforms",
     srcs = [
+        "AnnotateDispatches.cpp",
         "CaptureDispatchDynamicDims.cpp",
         "CleanupNumericNarrowing.cpp",
         "CleanupTensorShapes.cpp",
